refactor(experimental): add support for compiling to a new transaction type (#2389)

+++ packages/library-legacy-sham/src/public-key.ts
@@ -15,7 +15,7 @@ export class PublicKey<TAddress extends string = string> {
         return this.#address;
     }
     #getByteArray() {
-        return getAddressEncoder().encode(this.#address);
+        return getAddressEncoder().encode(this.#address) as Uint8Array;
     }
     #getBuffer() {
         if (!__NODEJS__ && typeof Buffer === 'undefined') {
